Reorganize gulp files
Major changes:
- move all config variables from gulp files to `gulp/config.js`
- move gulp task files to `gulp/tasks`
- move `doiuse` gulp task  to `gulp/tasks/check.js`

Other changes:
- rename some config variables
