+++ gulp/tasks/clean.js
