+++ gulp/tasks/docs.js
