+++ gulp/tasks/javascript.js
@@ -3,31 +3,13 @@ var concat = require('gulp-concat');
 var babel = require('gulp-babel');
 var onBabelError = require('./babel-error.js');
 
-var FOUNDATION = [
-  'js/foundation.core.js',
-  'js/foundation.util.*.js',
-  'js/*.js'
-];
-
-var DEPS = [
-  'node_modules/jquery/dist/jquery.js',
-  'node_modules/motion-ui/dist/motion-ui.js',
-  'node_modules/what-input/what-input.js'
-];
-
-var DOCS = [
-  'node_modules/clipboard/dist/clipboard.js',
-  'node_modules/corejs-typeahead/dist/typeahead.bundle.js',
-  'node_modules/foundation-docs/js/**/*.js',
-  'docs/assets/js/docs.*.js',
-  'docs/assets/js/docs.js'
-];
+var CONFIG = require('../config.js');
 
 // Compiles JavaScript into a single file
 gulp.task('javascript', ['javascript:foundation', 'javascript:deps', 'javascript:docs']);
 
 gulp.task('javascript:foundation', function() {
-  return gulp.src(FOUNDATION)
+  return gulp.src(CONFIG.JS_FILES)
     .pipe(babel()
       .on('error', onBabelError))
     .pipe(gulp.dest('_build/assets/js/plugins'))
@@ -36,13 +18,13 @@ gulp.task('javascript:foundation', function() {
 });
 
 gulp.task('javascript:deps', function() {
-  return gulp.src(DEPS)
+  return gulp.src(CONFIG.JS_DEPS)
     .pipe(concat('vendor.js'))
     .pipe(gulp.dest('_build/assets/js'));
 });
 
 gulp.task('javascript:docs', function() {
-  return gulp.src(DOCS)
+  return gulp.src(CONFIG.JS_DOCS)
     .pipe(concat('docs.js'))
     .pipe(gulp.dest('_build/assets/js'));
 });
